// src/app/api/model/tvSeasonImages.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Image } from './image';


export interface TvSeasonImages { 
    id?: number;
    posters?: Array<Image>;
}